Take a stab at documenting the SIG Roadmap
This commit creates the initial version of the SIG Roadmap
based on the recent conversations that took place during the
SIG Software Supply Chain Meetings.

The roadmap is not exhaustive and should be iterated further
with the contributions from the community.

This roadmap is heavy influenced by the SIG Interoperability
roadmap.